requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
Flask-SeaSurf==2.0.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
